refactor(backend): split backend/mod.rs into submodules + release 3.2.1 (MIK-6863)
Split the 3038-line src/backend/mod.rs into pool/lifecycle/ops/metadata/cached_metadata/registry/annotations submodules (+ tests/pool_tests), each under the 800-LOC hygiene cap. mod.rs is now a 73-line module root; all public paths (BackendRegistry, normalize_tool_annotations, runtime_plan_for_backend) preserved. Pure move refactor: no behaviour change, no public-API change, identical test coverage. fmt + clippy -D warnings + full suite green. Closes the MIK-6735 3.2.0 DoD deviation.

Local-Tested: cargo fmt+clippy+test green @ aaa51f912ef81aa0c18f08a4e326a2ceaf87b925
