+++ src/backend/pool.rs
@@ -1,0 +1,260 @@
+//! Per-identity transport/session pool (MIK-6735): [`PoolKey`], [`PooledEntry`],
+//! and the [`super::Backend`] methods that create, look up, and idle-evict
+//! pool slots.
+
+use std::sync::Arc;
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::time::Duration;
+
+use parking_lot::RwLock;
+use tokio::sync::Mutex;
+
+use super::Backend;
+use crate::failsafe::Failsafe;
+use crate::transport::Transport;
+
+/// Seconds since the Unix epoch, saturating to 0 on a pre-epoch clock.
+pub(crate) fn now_unix_secs() -> u64 {
+    std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs()
+}
+
+/// Identifies one transport/session slot in a backend's connection pool
+/// (MIK-6735).
+///
+/// A backend always owns the canonical [`PoolKey::Shared`] slot — the
+/// single-tenant default that also backs init, metadata, and canonical traffic.
+/// When `identity_propagation.session_mode = per_user` is configured and a
+/// caller identity is present, the backend additionally owns one
+/// [`PoolKey::PerUser`] slot per stable identity binding, so two distinct users
+/// never share a backend transport or its upstream MCP session (IDP.7).
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub(crate) enum PoolKey {
+    /// The canonical single-tenant slot. Every non-per-user backend, and every
+    /// per-user backend request that lacks a resolved identity, collapses here
+    /// so single-tenant behavior is preserved byte-for-byte (IDP.5).
+    Shared,
+    /// A per-user slot keyed by the caller's stable identity binding
+    /// (`PropagatedCredential::cache_binding`, MIK-6784).
+    PerUser { binding: String },
+}
+
+/// One pooled transport slot: its lazily started transport, a start lock that
+/// serializes connection setup for that slot, a last-used clock driving idle
+/// eviction of per-user slots, and this slot's own failsafe mechanisms.
+///
+/// The failsafe (circuit breaker + rate limiter + retry policy + health
+/// tracker) is owned per-slot, not per-backend (MIK-6735 fix 1, adversarial
+/// review of commit bfd62b91). Gating `request_with_headers` on a single
+/// backend-wide `Failsafe` meant one caller identity's transport failing
+/// enough tripped the breaker for every OTHER identity sharing the same
+/// backend too — the exact cross-tenant blast radius the per-user pool
+/// exists to eliminate. Each slot now fails independently: the Shared slot
+/// keeps its own failsafe (behavior for non-per-user backends is byte-for-
+/// byte unchanged), and each `PerUser` slot gets a fresh one the moment it is
+/// first created.
+pub(crate) struct PooledEntry {
+    pub(crate) transport: RwLock<Option<Arc<dyn Transport>>>,
+    pub(crate) start_lock: Mutex<()>,
+    pub(crate) last_used: AtomicU64,
+    pub(crate) failsafe: Failsafe,
+}
+
+impl PooledEntry {
+    pub(crate) fn new(name: &str, failsafe_config: &crate::config::FailsafeConfig) -> Self {
+        Self {
+            transport: RwLock::new(None),
+            start_lock: Mutex::new(()),
+            last_used: AtomicU64::new(now_unix_secs()),
+            failsafe: Failsafe::new(name, failsafe_config),
+        }
+    }
+
+    /// Mark this slot as used now, deferring its idle eviction.
+    pub(crate) fn touch(&self) {
+        self.last_used.store(now_unix_secs(), Ordering::Relaxed);
+    }
+}
+
+impl Backend {
+    /// The backend's configured session mode, if identity propagation is set.
+    pub(super) fn session_mode(&self) -> Option<crate::identity_propagation::SessionMode> {
+        self.config
+            .identity_propagation
+            .as_ref()
+            .map(|c| c.session_mode)
+    }
+
+    /// Derive the pool slot for a request carrying `identity_key`.
+    ///
+    /// Only a `per_user` backend with a concrete caller identity gets its own
+    /// slot; every other case — no identity propagation, `stateless`, or
+    /// `per_user` without a resolved identity — collapses to the shared
+    /// canonical slot, preserving single-tenant behavior byte-for-byte (IDP.5).
+    pub(super) fn pool_key_for(&self, identity_key: Option<&str>) -> PoolKey {
+        use crate::identity_propagation::SessionMode;
+        match (self.session_mode(), identity_key) {
+            (Some(SessionMode::PerUser), Some(binding)) => PoolKey::PerUser {
+                binding: binding.to_string(),
+            },
+            _ => PoolKey::Shared,
+        }
+    }
+
+    /// Fetch (or lazily create) the pooled entry for `key`. The `Arc` is cloned
+    /// out so the `DashMap` shard guard is released before any `.await`.
+    ///
+    /// Logs + gauges the live slot count on creation only (MIK-6735 fix 3) —
+    /// minimal observability into per-user pool growth without a per-request
+    /// cost on the (overwhelmingly more common) cache-hit path.
+    pub(super) fn pooled_entry(&self, key: &PoolKey) -> Arc<PooledEntry> {
+        let mut created = false;
+        let entry = Arc::clone(
+            self.pool
+                .entry(key.clone())
+                .or_insert_with(|| {
+                    created = true;
+                    Arc::new(PooledEntry::new(&self.name, &self.failsafe_config))
+                })
+                .value(),
+        );
+        if created {
+            #[allow(clippy::cast_precision_loss)] // pool size is never remotely close to 2^52
+            let live = self.pool.len() as f64;
+            telemetry_metrics::gauge!(
+                "mcp_backend_pool_slots",
+                "backend" => self.name.clone()
+            )
+            .set(live);
+            tracing::debug!(backend = %self.name, ?key, live_slots = live, "Pool slot created");
+        }
+        entry
+    }
+
+    /// The canonical shared slot's `PooledEntry`. Inserted at construction and
+    /// never evicted (`evict_idle_per_user_entries` explicitly skips it), so
+    /// this is always present — used by status/metrics/health-loop accessors
+    /// that intentionally report the backend-wide, single-tenant view
+    /// regardless of how many per-user slots exist (MIK-6735 fix 1).
+    pub(super) fn shared_entry(&self) -> Arc<PooledEntry> {
+        Arc::clone(
+            self.pool
+                .get(&PoolKey::Shared)
+                .expect("PoolKey::Shared is inserted at construction and never evicted")
+                .value(),
+        )
+    }
+
+    /// Clone the canonical shared slot's live transport, if started.
+    pub(super) fn shared_transport(&self) -> Option<Arc<dyn Transport>> {
+        self.pool
+            .get(&PoolKey::Shared)
+            .and_then(|entry| entry.value().transport.read().clone())
+    }
+
+    /// Idle-evict per-user pool slots whose last use predates `idle_ttl`,
+    /// closing their transports. The canonical [`PoolKey::Shared`] slot is never
+    /// evicted (it backs init, metadata, and single-tenant traffic). Returns the
+    /// number of slots closed (MIK-6735 POOL.2).
+    pub async fn evict_idle_per_user_entries(&self, idle_ttl: Duration) -> usize {
+        let cutoff = idle_ttl.as_secs();
+
+        // First pass: collect candidate keys without holding a guard across the
+        // async close(). Skip the shared slot outright.
+        let candidates: Vec<PoolKey> = self
+            .pool
+            .iter()
+            .filter(|entry| !matches!(entry.key(), PoolKey::Shared))
+            .map(|entry| entry.key().clone())
+            .collect();
+
+        let mut closed = 0;
+        for key in candidates {
+            // Atomically remove only if STILL idle — re-checked inside the shard
+            // lock so a request that touched the slot after the first pass keeps
+            // it alive and is never torn down mid-flight.
+            let removed = self.pool.remove_if(&key, |k, entry| {
+                !matches!(k, PoolKey::Shared)
+                    && now_unix_secs().saturating_sub(entry.last_used.load(Ordering::Relaxed))
+                        >= cutoff
+            });
+            if let Some((_, entry)) = removed {
+                let transport = entry.transport.write().take();
+                if let Some(transport) = transport {
+                    let _ = transport.close().await;
+                }
+                closed += 1;
+            }
+        }
+        if closed > 0 {
+            // MIK-6735 fix 3: gauge + log the live slot count after eviction,
+            // mirroring the creation-side observability in `pooled_entry`.
+            #[allow(clippy::cast_precision_loss)] // pool size is never remotely close to 2^52
+            let live = self.pool.len() as f64;
+            telemetry_metrics::gauge!(
+                "mcp_backend_pool_slots",
+                "backend" => self.name.clone()
+            )
+            .set(live);
+            tracing::debug!(
+                backend = %self.name,
+                evicted = closed,
+                live_slots = live,
+                "Idle per-user pool slots evicted"
+            );
+        }
+        closed
+    }
+
+    #[cfg(test)]
+    pub(crate) fn set_transport_for_test(&self, transport: Arc<dyn Transport>) {
+        let entry = self.pooled_entry(&PoolKey::Shared);
+        *entry.transport.write() = Some(transport);
+    }
+
+    /// Test-only: inject a transport into a specific pool slot so isolation
+    /// tests can seed distinct per-user sessions (MIK-6735 POOL.4).
+    #[cfg(test)]
+    pub(crate) fn set_pooled_transport_for_test(
+        &self,
+        key: &PoolKey,
+        transport: Arc<dyn Transport>,
+    ) {
+        let entry = self.pooled_entry(key);
+        *entry.transport.write() = Some(transport);
+    }
+
+    /// Test-only: clone the transport `Arc` stored in a specific pool slot, so
+    /// isolation tests can assert distinct instances via `Arc::ptr_eq`.
+    #[cfg(test)]
+    pub(crate) fn pooled_transport_for_test(&self, key: &PoolKey) -> Option<Arc<dyn Transport>> {
+        self.pool
+            .get(key)
+            .and_then(|entry| entry.value().transport.read().clone())
+    }
+
+    /// Test-only: trip this backend's canonical Shared-slot circuit breaker
+    /// open by recording `failure_threshold` consecutive failures.
+    #[cfg(test)]
+    pub(crate) fn trip_circuit_breaker_for_test(&self) {
+        self.trip_circuit_breaker_for_test_key(&PoolKey::Shared);
+    }
+
+    /// Test-only: trip an arbitrary pool slot's circuit breaker open
+    /// (MIK-6735 fix 1) — generalizes [`Self::trip_circuit_breaker_for_test`]
+    /// (Shared-only) to any [`PoolKey`], so cross-tenant isolation tests can
+    /// trip one identity's slot without touching another's.
+    #[cfg(test)]
+    pub(crate) fn trip_circuit_breaker_for_test_key(&self, key: &PoolKey) {
+        let entry = self.pooled_entry(key);
+        let threshold = entry.failsafe.circuit_breaker.stats().failure_threshold;
+        for _ in 0..threshold {
+            entry
+                .failsafe
+                .circuit_breaker
+                .record_failure("test-trip", std::time::Duration::ZERO);
+        }
+    }
+}
